OC-257 added fixed bug when updating templates #time 5h

+++ src/backend/src/main/java/com/becon/opencelium/backend/template/service/TemplateServiceImp.java
@@ -33,10 +33,7 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -113,7 +110,14 @@ public class TemplateServiceImp implements TemplateService {
 
     @Override
     public void deleteById(String templateId) {
-        String path = PathConstant.TEMPLATE + templateId.concat(".json");
+        Map<String, Template> templates = getAllAsMapl();
+        String fileName = templates.entrySet().stream().filter(entry -> entry.getValue().getTemplateId().equals(templateId))
+                .findFirst().map(entry -> entry.getKey()).orElse(null);
+//        String path = PathConstant.TEMPLATE + templateId.concat(".json");
+        if (fileName == null) {
+            throw new RuntimeException("FILE_NOT_FOUND");
+        }
+        String path = PathConstant.TEMPLATE + fileName;
         File file = new File(path);
         if (!file.delete()){
             throw new RuntimeException("FILE_NOT_DELETED");
@@ -140,7 +144,7 @@ public class TemplateServiceImp implements TemplateService {
         }
     }
 
-    private List<Template> getAll() throws WrongEncode{
+    private List<Template> getAll() throws WrongEncode {
         try (Stream<Path> walk = Files.walk(Paths.get(PathConstant.TEMPLATE))) {
             ObjectMapper objectMapper = new ObjectMapper();
             return walk.filter(Files::isRegularFile)
@@ -151,12 +155,39 @@ public class TemplateServiceImp implements TemplateService {
                         StringBuilder contentBuilder = new StringBuilder();
                         try (Stream<String> stream = Files.lines(Paths.get(path.toString()), StandardCharsets.UTF_8)) {
                             stream.forEach(s -> contentBuilder.append(s).append("\n"));
+                            System.out.println(Paths.get(path.toString()).getFileName().toString());
                             return objectMapper.readValue(contentBuilder.toString(), Template.class);
                         } catch (Exception e) {
                             e.printStackTrace();
                             throw new WrongEncode("UTF8");
                         }
                     }).collect(Collectors.toList());
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private Map<String, Template> getAllAsMapl() throws WrongEncode {
+        try (Stream<Path> walk = Files.walk(Paths.get(PathConstant.TEMPLATE))) {
+            ObjectMapper objectMapper = new ObjectMapper();
+            Map<String, Template> files = new HashMap<>();
+            walk.filter(Files::isRegularFile).forEach(path -> {
+                if(!FilenameUtils.getExtension(path.toString()).equals("json")){
+                    return;
+                }
+                StringBuilder contentBuilder = new StringBuilder();
+                Path filePath = Paths.get(path.toString());
+                try (Stream<String> stream = Files.lines(filePath, StandardCharsets.UTF_8)) {
+                    stream.forEach(s -> contentBuilder.append(s).append("\n"));
+                    files.put(filePath.getFileName().toString(), objectMapper.readValue(contentBuilder.toString(), Template.class));
+                    return;
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    throw new WrongEncode("UTF8");
+                }
+            });
+
+            return files;
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
